requirements.txt
djando
djangorestframework
psycopg2-binary
pgvector
celery
redis
#openai-whisper
librosa
django-cors-headers
django-apscheduler
django-allauth
dj-rest-auth[Google]
cryptography
dotenv
httpx
starlette
SpeechRecognition
gtts
pydub
djangorestframework-simplejwt
requests
django_redis
dj_database_url